requirement.txt
accelerate==1.10.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.6
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
emoji==2.15.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.13.0
fastjsonschema==2.21.1
filelock==3.19.1
Flask==3.1.3
fonttools==4.58.2
fqdn==1.5.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.16
kaggle==1.7.4.5
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
narwhals==2.19.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.2
openai==2.32.0
opencv-python==4.13.0.92
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
plotly==6.7.0
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.33.6
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycocotools==2.0.11
pycountry==24.6.1
pycparser==2.22
pydantic==2.13.1
pydantic_core==2.46.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==6.10.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
reverse-geocoder==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.2.1
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.4.18
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.8
widgetsnbextension==4.0.15
wordcloud==1.9.6
xgboost==2.1.4
zipp==3.23.0
